#1 BasicImplementation - final adjustments / cleanups
